Update versions in package.json
This commit updates the versions for libraries and minimum
Node in package.json. Technically, the ldapjs update to 1.0.0
is only required if you want to run the tests, but there's no
sense in not using it for other things.

The minimum Node version was verified against the compatibility
table -- https://kangax.github.io/compat-table/es6/