Really and truly fixes the paged search results
Paging was still broken. It was only returning a single page of
results where a page size was defined by the remote AD server.
This implementation gets ALL of the results. In the process,
issue #122 is resolved.

Note: this commit locks the bunyan dependency to 1.3.5. Newer
versions of this library cause the Node debugger to segfault.
That made resolving this paging issue quite fun.