Establish fork so it can be published to npmjs.com
Maybe upstream will merge us in one day, but I can't wait
any longer. I need to be able to define this as a publicly
accessible dependency that can be acquired through an
npm proxy (i.e. npm-lazy).

I am establishing this fork as version 1.0.0 since it is
a major overhaul of the whole library.